Criacao de regra que limita a exclusao de um Coordenador vinculado a um Historico de Coordenadores

+++ src/java/nst/controlenst/controller/business/impl/CoordenadorBO.java
@@ -11,6 +11,7 @@ import java.util.ArrayList;
 import nst.controlenst.controller.business.exception.BusinessExceptions;
 import nst.controlenst.model.entity.Coordenador;
 import nst.controlenst.persistence.dao.factory.interfaces.CoordenadorDAO;
+import nst.controlenst.persistence.dao.factory.interfaces.HistoricoCoordenadorDAO;
 import nst.controlenst.persistence.dao.util.FabricaDAO;
 
 /**
@@ -20,11 +21,13 @@ import nst.controlenst.persistence.dao.util.FabricaDAO;
 public class CoordenadorBO implements IBusiness {
 
     private CoordenadorDAO coordenadorDAO = null;
+    private HistoricoCoordenadorDAO historicoCoordenadorDAO = null;
     private Coordenador coordenador = null;
 
     public CoordenadorBO() {
         try {
             this.coordenadorDAO = FabricaDAO.getFactoryType().getCoordenadorDAO();
+            this.historicoCoordenadorDAO = FabricaDAO.getFactoryType().getHistoricoCoordenadorDAO();
         } catch (Exception ex) {
             Logger.getLogger(CoordenadorBO.class.getName()).log(Level.SEVERE, null, ex);
         }
@@ -34,11 +37,27 @@ public class CoordenadorBO implements IBusiness {
     public void excluir(Object objeto) throws BusinessExceptions {
         this.coordenador = (Coordenador) objeto;
 
+        /*
+         * Nao pode ter Id nulo ou Zero
+         */
         if (this.coordenador.getId() == null || this.coordenador.getId() == 0) {
             throw new BusinessExceptions("Não foi possível identificar o valor do índice no objeto.");
         }
-
+        
+        /*
+         * Nao permite a exclusao de um Coordenador que esteja vinculado a um registro no Historico.
+         */
+        if(!this.historicoCoordenadorDAO.getAllByCoordenador(this.coordenador).isEmpty()){
+            throw new BusinessExceptions("Não é permitido a exclusão de um Coordenador vinculado a um registro de Historico de Coordenadores.");
+        }
+        
         this.coordenadorDAO.delete(coordenador);
+        
+        
+        
+        //Não permitir a exclusao de um Coordenador que esteja vinculado a um projeto.
+        
+        
     }
 
     @Override
@@ -55,5 +74,5 @@ public class CoordenadorBO implements IBusiness {
         }
 
         this.coordenadorDAO.save(coordenador);
-    }
+    }    
 }
